+++ pkg/account/downloader/internal/http/http_utils.go
@@ -21,23 +21,7 @@ import (
 )
 
 func closeResponseBody(resp *http.Response) {
-	_ = resp.Body.Close()
-}
-
-func handleClientResponseError(resp *http.Response, clientErr error) error {
-	return clientErr
-
-	//TODO: change that 404 not found return nil. Needs code adaptation
-	//if clientErr != nil && resp == nil {
-	//	return clientErr
-	//}
-	//
-	//if !rest.IsSuccess(resp) && resp.StatusCode != http.StatusNotFound {
-	//	body, err := io.ReadAll(resp.Body)
-	//	if err != nil {
-	//		return fmt.Errorf("unable to read response body %w", err)
-	//	}
-	//	return fmt.Errorf("(HTTP %d): %s", resp.StatusCode, string(body))
-	//}
-	//return nil
+	if resp != nil {
+		_ = resp.Body.Close()
+	}
 }
